extract common service code for re use in other HTTP sinks
